--- modules/cms-kit/host/Volo.CmsKit.Web.Unified/ViewComponents/CommentDate.cshtml
@@ -1,11 +1,0 @@
-﻿@model Volo.CmsKit.ViewComponents.CommentDateViewComponent
-
-<p>Welcome Comment Date Component</p>
-@if (Model.IsShow)
-{
-    <p> @DateTime.Today.ToLongDateString()</p>
-}
-else
-{
-    <p>Without date</p>
-}
