--- modules/cms-kit/host/Volo.CmsKit.Web.Unified/ViewComponents/CommentDateViewComponent.cs
@@ -1,25 +1,0 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc.UI.Widgets;
-
-namespace Volo.CmsKit.ViewComponents;
-
-[Widget(
-    AutoInitialize = true
-)]
-
-[ViewComponent(Name = "CommentDate")]
-public class CommentDateViewComponent : AbpViewComponent
-{
-    public bool IsShow { get; set; }
-
-    public CommentDateViewComponent()
-    {
-    }
-
-    public virtual async Task<IViewComponentResult> InvokeAsync(string isShow)
-    {
-        return View("~/ViewComponents/CommentDate.cshtml", new CommentDateViewComponent() { IsShow = bool.Parse(isShow) });
-    }
-}
