+++ modules/cms-kit/host/Volo.CmsKit.Web.Unified/ViewComponents/Format.cshtml
@@ -2,5 +2,5 @@
 @model DecisionCommentDateViewModel
 
 <div class="form-check mb-3">
-    <abp-input asp-for="IsShow" />
+    <abp-input asp-for="Format" />
 </div>
