+++ modules/cms-kit/host/Volo.CmsKit.Web.Unified/ViewComponents/Today.cshtml
@@ -1,0 +1,5 @@
+﻿@model Volo.CmsKit.ViewComponents.TodayViewComponent
+
+<p>Welcome Today Component</p>
+
+<p> @DateTime.Now.ToString(Model.Format)</p>
