+++ modules/cms-kit/src/Volo.CmsKit.Common.Application/Volo/CmsKit/Contents/CmsKitContentWidgetOptions.cs
@@ -11,7 +11,7 @@ public class CmsKitContentWidgetOptions
         WidgetConfigs = new();
     }
 
-    public void AddWidget(string widgetKey, string widgetName, string editorWidgetName)
+    public void AddWidget(string widgetKey, string widgetName, string editorWidgetName = null)
     {
         var config = new ContentWidgetConfig(widgetName, editorWidgetName);
         WidgetConfigs.Add(widgetKey, config);
